+++ examples/js/manipulation/demo.js
@@ -4,6 +4,7 @@ import DeleteRowTable from './delete-row-table';
 import SearchTable from './search-table';
 import ColumnFilterTable from './filter-table';
 import MultiSearchTable from './multi-search-table';
+import ExportCSVTable from './export-csv-table';
 
 class Demo extends React.Component {
   render() {
@@ -29,9 +30,10 @@ class Demo extends React.Component {
         </div>
         <div className="col-md-offset-1 col-md-8">
           <div className="panel panel-default">
-            <div className="panel-heading">Table Search Example</div>
+            <div className="panel-heading">Table Search Example(Include after search hook)</div>
             <div className="panel-body">
               <h5>Source in /examples/js/manipulation/search-table.js</h5>
+              <h5>See your console to watch information of after searching.</h5>
               <SearchTable />
             </div>
           </div>
@@ -50,7 +52,17 @@ class Demo extends React.Component {
             <div className="panel-heading">Column Filter Example</div>
             <div className="panel-body">
               <h5>Source in /examples/js/manipulation/filter-table.js</h5>
+              <h5>See your console to watch information of after column filtering.</h5>
               <ColumnFilterTable />
+            </div>
+          </div>
+        </div>
+        <div className="col-md-offset-1 col-md-8">
+          <div className="panel panel-default">
+            <div className="panel-heading">Export CSV Example</div>
+            <div className="panel-body">
+              <h5>Source in /examples/js/manipulation/export-csv-table.js</h5>
+              <ExportCSVTable />
             </div>
           </div>
         </div>